# serverless/functions/src/router/crypto/coins.enums.ts
export enum Tag {
  'DAO' = 'dao',
  'DEFI' = 'defi',
  'DEX' = 'dex',
  'EXCHANGE' = 'exchange',
  'GAMING' = 'gaming',
  'LAYER_1' = 'layer-1',
  'LAYER_2' = 'layer-2',
  'MEME' = 'meme',
  'METAVERSE' = 'metaverse',
  'NFT' = 'nft',
  'PRIVACY' = 'privacy',
  'STABLECOIN' = 'stablecoin',
  'STAKING' = 'staking',
  'WRAPPED' = 'wrapped',
}

export enum TimePeriod {
  'HOUR_1' = '1h',
  'HOUR_3' = '3h',
  'HOUR_12' = '12h',
  'HOUR_24' = '24h',
  'DAY_7' = '7d',
  'DAY_30' = '30d',
  'MONTH_3' = '3m',
  'YEAR_1' = '1y',
  'YEAR_3' = '3y',
  'YEAR_5' = '5y',
}

export enum Tier {
  'TIER_1' = '1',
  'TIER_2' = '2',
  'TIER_3' = '3',
}

export enum OrderBy {
  '24H_VOLUME' = '24hVolume',
  'CHANGE' = 'change',
  'LISTED_AT' = 'listedAt',
  'MARKET_CAP' = 'marketCap',
  'PRICE' = 'price',
}

export enum OrderDirection {
  'ASC' = 'asc',
  'DESC' = 'desc',
}